+++ prost-reflect-tests/src/json.rs
@@ -946,7 +946,7 @@ where
         options,
     )
     .unwrap()
-    .to_message()
+    .transcode_to()
     .unwrap()
 }